requirements.txt
Flask
python-docx
textdistance
beautifulsoup4
gunicorn
